Move EnsureIndexes to elastic search package.
